add astropy_interface
